v10: TMP FIX for template variables
The use of `t-call` tags with variables containing template names is broken.

So far we used `t-call="#{ form.form_search_results_template }"` and alike
to render custom templates on the fly. This is not working anymore in v10.

A quick debugging session showed that `QWeb._compile_directive_call`
does not evaluate the content of `t-call`, it  uses it as a template name.